+++ client/src/pages/Test/components/SearchList/SearchList.module.css
@@ -1,0 +1,44 @@
+.filterSection {
+  display: flex;
+
+  gap: 5rem;
+}
+
+.filterLayout {
+  display: flex;
+
+  position: relative;
+  z-index: 999;
+}
+
+.filterLabel {
+  font-size: 20px;
+
+  cursor: pointer;
+}
+
+.dropdownList {
+  display: flex;
+  flex-direction: column;
+
+  position: absolute;
+  top: 2.7rem;
+
+  border-radius: 2px;
+  box-shadow: 0 0 2px;
+
+  background-color: #ffffff;
+}
+
+.dropdownList li {
+  width: 10rem;
+  padding: 0.5rem;
+
+  font-size: 12px;
+
+  cursor: pointer;
+}
+
+.dropdownList li:hover {
+  background: #efefef;
+}
